requirements.txt
Flask
requests
debugpy # Required for debugging.
google-cloud-aiplatform